requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
firebase-admin==6.3.0
gunicorn==21.2.0
cachelib==0.10.2
